updated 10-main.c

+++ 0x02-functions_nested_loops/10-add.c
@@ -2,13 +2,16 @@
 
 /**
  * add - adds two integer and returns the result
+ *
  * @a: first number
  * @b: second number
+ *
  * Return: result
  */
+
 int add(int a, int b)
 {
-	int r;
+	long r;
 
 	r = a + b;
 	return (r);
